# Backend/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Keep-Track API
  description: Keep-Track API
  version: 1.0.0

paths:
  # product controller
  /product:
    get:
      tags:
        - Product
      operationId: getProduct
      description: Returns the Product for the given barcode
      parameters:
        - name: barCode
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '500':
          description: Server error
    post:
      tags:
        - Product
      operationId: postProduct
      description: Adds new product to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Successfully added new product
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Server error

  # storage controller
  /storages:
    get:
      tags:
        - Storage
      operationId: getStorageOverviewsOfUser
      description: Returns all storage overviews of given user
      responses:
        '200':
          description: Retrieved all storages for given user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StorageOverview'
        '500':
          description: Server error
  /storage/{storageId}:
    get:
      tags:
        - Storage
      operationId: getStorageByStorageId
      description: Returns specific storage for storage id
      parameters:
        - name: storageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved storage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Storage'
        '500':
          description: Server error

  # user controller
  /user/register:
    post:
      tags:
        - User
      operationId: addUser
      description: Registers new User to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User successfully registered
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Server error
  /user/register/confirm:
    get:
      tags:
        - User
      operationId: confirmRegistration
      description: Confirm a pending user registration
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User successfully confirmed
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Server error


components:
  schemas:
    Product:
      type: object
      properties:
        productName:
          type: string
        manufacturer:
          type: string
        barCode:
          type: string
    Storage:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        owner:
          type: integer
          format: int64
        products:
          type: array
          items:
            $ref: '#/components/schemas/StorageEntry'
    StorageOverview:
      type: object
      properties:
        name:
          type: string
        owner:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
    StorageEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        storageId:
          type: integer
          format: int64
        product:
          $ref: '#/components/schemas/Product'
        amount:
          type: integer

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        jwt:
          type: string
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
        - surname
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        surname:
          type: string